+++ CHANGELOG.md
@@ -9,6 +9,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/)
 
 ### Changed
 
+- The Entity table in Spine Database editor can now be properly filtered by selecting scenarios
+  from the Scenario tree.
+  When doing so, the table will show the entities that are visible in the selected scenario.
+
 ### Deprecated
 
 ### Removed